refactor: NeteaseCloudMusicConfig字段类型修改
phoneNumber Int 2 Long

+++ model/src/main/kotlin/ski/mashiro/entity/config/NeteaseCloudMusicConfig.kt
@@ -4,7 +4,7 @@ package ski.mashiro.entity.config
  * @author mashirot
  */
 data class NeteaseCloudMusicConfig(
-    var phoneNumber: Int?,
+    var phoneNumber: Long?,
     val password: String?,
     var passwordMD5: String?,
     var cookie: String,
